Changed functionality of controller - remove controller GetCardsServlet
