--- src/main/java/com/greetingcard/web/controller/GetCardsController.java
@@ -1,47 +1,0 @@
-package com.greetingcard.web.controller;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.greetingcard.entity.Card;
-import com.greetingcard.entity.User;
-import com.greetingcard.service.CardService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.HttpSession;
-import java.util.List;
-import java.util.Map;
-
-@Slf4j
-@RestController
-@RequestMapping(value = "/api/v1", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-public class GetCardsController {
-    private CardService cardService;
-
-    @Autowired
-    public GetCardsController(CardService cardService) {
-        this.cardService = cardService;
-    }
-
-    @GetMapping(value = "/cards")
-    public ResponseEntity<Object> getCards(HttpSession session, @RequestParam String type) throws JsonProcessingException {
-        log.info("getCards");
-        User user = (User) session.getAttribute("user");
-        long userId = user.getId();
-        List<Card> cardList = cardService.getCards(userId, type);
-        if (cardList == null) {
-            ObjectMapper mapper = new ObjectMapper();
-            String json = mapper.writeValueAsString(Map.of("message", "Sorry, you do not have cards"));
-            return ResponseEntity.status(HttpStatus.OK).body(json);
-        } else {
-            return ResponseEntity.status(HttpStatus.OK).body(cardList);
-        }
-    }
-}
